Add debug prints for playback and decoder
The decoder is currently blocked for long times, and it looks like that
is because the playback thread is holding the lock. I suspect it has to
do with the write call blocking; the Alsa docs for writei say:

> If the blocking behaviour is selected and it is running, then routine
> waits until all requested frames are played or put to the playback
> ring buffer. The returned number of frames can be less only if a
> signal or underrun occurred.

I expected that in the common case it would put things in the buffer,
and that that would be fast, but now that I am thinking about it, that
behavior would not be terribly useful, so it makes sense that the call
blocks.

I can change to non-blocking mode, but then I got an EAGAIN on the drain
call. Handling those everywhere seems tedious too, so I am first going
to investigate if this mmap rather than writei is useful after all.