Simplify decode buffer size computation
If it's a strict maximum, then we have to be careful to stay under it,
but we are not going to stay under it anyway, because we will try again
next round in a new decode task if the buffer is not full. So instead,
just stop as soon as we went over the limit. The size of a flac frame is
not so large that we will go over the limit by a lot, and it simplifies
the semantics.