Install Alsa through Apt on CI after all
This feels like admitting defeat. I’m pretty sure it can be done with
Nix, but then maybe I need to move the entire build into Nix, which has
other problems. Besides, using Nix to manage build tools is fine, but if
you are not going to run the binary on NixOS or deploy it with Nix, then
you probably want it to link against the system Alsa, not against the
one in the Nix store. So maybe this is for the best after all.