Record some measurements for new disk IO experiment
I know this is only one sample, so I should not draw conclusions from
this, but it is pretty clearly a regression, and I have a suspicion why:
files that are close together on disk don't get spread out over threads
as much due to chunking. Let me change the feeding strategy, I expect
that will help.