Change load order to be interleaved
This way, the "frontier" of what threads are loading will remain in
enumeration order, it will not be that every thread has its own chunk
that it is racing on independently. My hypothesis is that files that are
close in the iteration order, are also close on disk, so this is a
better pattern to minimize seeks.