Simplify file reading to atomic counter into slice
I have a slice of paths anyway. Instead of putting them in a channel,
just let threads increment an atomic int, so they will take the next one
from the slice, and you get this "frontier" without the complexity.

This brings back the time for the CPU-bound case to the baseline, maybe
even slightly below it, but I did not measure enough to confirm that.

I just noticed that the size of the disk queue makes a huge difference,
so the stuff I imagined that fadvise would help is not going to make a
difference, I think. I just need to open enough threads to keep the IO
queues fed, but no more than that, because that would cause
fragmentation. That also explains the tuning.