Record performance after tuning disk nr_requests
It looks like this is a much bigger factor, and once that parameter is
cranked up, adding threads helps, but only marginally though.

For CPU-bound indexing though, we shouldn't have too many threads, so
there is a trade-off here.

I have an idea: let's start new threads dynamically. If we started one,
and find that little progress has been made in the meantime, then we
could start another thread. If it did make a lot of progress, then we
are probably CPU bound, so we can safely stop spawning more threads.