Bump Rust toolchain to 1.43.1
I want to update some of the packages, and most of the Rust ecosystem no
longer compiles with this older version, mainly due to the MaybeUninit
fiasco. So either I use older versions of packages, or I use a newer
Rust.

However, in case somebody wants to pick this up and package it, I don't
want to be a burden. I can imagine especially that it would be useful on
Raspbian, a Debian derivative, so it should at least play nicely with
Debian.

Currently Debian Jessie, Stretch, and Buster, all ship Rust 1.34.2, the
previously pinned version, but MaybeUninit needs at least 1.36. Debian
unstable however is up to date with the lastest Rust. Ubuntu is fairly
up to date too, it ships 1.43 even to 16.04. NixOS ships 1.44.1 at the
time of writing in the unstable channel. So I think going with 1.43
should be safe. In case somebody does want to package this, it will not
be backported anyway, so focussing on the current releases should be
fine.

Ubuntu LTS is the lowest common denominator here, so I will go with that
for now, 1.43 it is. Ubuntu ships 1.43.0, but 1.43.1 exists, which fixes
a security issue, so I'll take that one.