Replace Hyper with tiny-http as the http server
That went smoother than I expected, the structure of the code did not
have to change a lot. This is a big reduction in dependencies, and I am
more confident now that I will control the server in the future, rather
than the server shoehorning me into structuring my code in weird ways.

It is still a blocking loop with a single thread, but that should be
easy to change in just a few lines. For now, this swaps out the server!

I also upgraded rust-toolchain because I wanted to use
str::strip_prefix, and given that all of the distros that I investigated
just yet are shipping a recent (but not always the most recent) compiler
anyway, I think that over time this will be fine.