Record absolute times in queue
For playback, it is useful to return the times as relative from the
instant the request was made; if we used absolute times, we need to
worry about synchronizing clocks between the server and client.

But as soon as the data is at the client, it is useful to convert it
into absolute times, because absolute times don't change all the time.
We can compute the current playback position from the start time.

Ideally it would use a monotonic clock at the client, but I don't think
such a thing exists in js, given how rudimentary the time support that
does exist is.