Reload queue after track change, and periodically
The queue is a dynamic thing, we can fetch it once, but it may change
under our feet. Fortunately the queue includes all the details necessary
to predict how it will change, so even with a single fetch, we could do
a good job ... if we are the only client, and if there are no bugs or
other problems that cause us to go out of sync. So therefore, reload the
queue periodically. That also means that for now there is no need to
predict how it will change, as long as we can predict *when* it will
change.