Refactor status bar to have one node per track
This way I can animate that entire node for fading it in/out, and it
also neatly sidesteps the problem of the progress bar reverting, because
there is one progress bar per track, so the next track will have a clear
new progress bar that starts at 0.

Transition animation has not been implemented yet, so right now the
statuses keep piling up.