Fix initial progress indicator value
Manipulating css from js is a tricky business. It seems that
modifications don't get applied immediately, but that there is some kind
of transaction, and all the modifications made in the same frame get
applied together, and that makes it difficult to reason about which
transition applies. I suppose Chromium does it for performance reasons,
recomputing layout is expensive, so you can do it once after all of the
js runs, instead of computing all intermediate steps that will not be
used anyway.

Adding a delay works, but this implementation is not ideal, because it
does not serialize all operations. So now two updates could race to
apply their modifications, such that the transition is set twice, and
then the transform twice, instead of having them interleaved. But the
time scale of the delay should be small enough, and I rate limit
progress updates, so I think in practice this will work.