Rename the project to "Musium"
I renamed it from the working title "Metaindex" to "Mindec" in a8ac829f,
but I was never really satisfied with the name. Now that I pivoted from
just indexing to being a fully player, the name is also not as
appropriate any more.

It is hard to come up with a good name that is nice but also
distinctive, I played a lot with Namium and Portmanteaur over the past
few weeks but counldn't find something much better than Mindec. Then
today I discovered the option "Musium", a contraction of "Music" and the
-ium suffic that Namium also uses.

A quick Google and DuckDuckGo search for "Musium" bring up nothing
interesting, mostly misspellings of "museum". A search for
"site:github.com 'musium'" does bring up a few hits, but all projects
look unrelated or abandoned. A search on GitHub itself does bring up a
few more repositories that are all empty or abandoned, so I think the
name is sufficiently distinctive.

What is nice is that it's 6 letters like Mindec, so all of the monospace
alignment can stay the same. The rename was automated though Sed.