Add a configuration module
This is a bit nicer to use than having to provide everything in the
right order on the command line. I considered using toml, but that adds
a nontrivial amount of dependencies. Serde derive is nice, but also
notorious for its build time. Just splitting on newlines and equals
gives me what I need, plus the potential for reporting better errors in
just a few lines of code.