Avoid usize overflow on ARM in position computation
This was an intersting one! When playing on my Raspberry Pi, I noticed
the progress bar would go backwards at times, while debugging on my
laptop this never happened.

Inspecting further, it was really the API serving the times that went
backwards. The current position would grow up to 100 seconds, and then
wrap around to 0 seconds. I suspected this had something to do with the
format string, maybe I misunderstood the format specifier, and I was
truncating something. But after some debugging, it turned out that the
integer number of milliseconds was wrapping!

The code didn’t show anything suspicious, surely a multiply by 500 would
not overflow a usize, even if a usize was 32-bit? Baffled I tried to
make a minimal example, which the compiler rejected due to overflowing
literals. It turns out, for only a few minutes of audio, multiplying the
sample count by 500 will already overflow u32 after all.

And consider this:

 * 100 seconds of stereo audio at 44.1 kHz is 8,820,000. Multiplied by
   500 we get 4,410,000,000.

 * 2^32 = 4,294,967,296.

So the overflow happens almost exactly at 100 seconds, which was why the
value was so misleading. It has nothing to do with decimals!

If common sample rates were all multiples of 500, we could divide the
denominator by 500 instead of multiplying the numerator, but
unfortunately 44100 Hz is not a multiple of 500, and it is the most
common sample rate. So move to 64-bit integers instead.