Implement queue ids
The act of queueing a track now assigns a "queue id" to that
enqueuement. This is a unique identifier for a track that was, is being,
or is going to be played at some point. The track id itself is
insufficient for this, because you can play, and even equeue the same
track many times in a row. The track id + queue position is also no
good, because the queue position changes, so it is difficult to use that
to refer to the same thing. A "queue id" gives queued tracks an identiy.

This can now be used to:

 * Ensure that the current track in the library browser updates, even
   when you enqueue the same track twice in a row.

 * Log when tracks were played, without worrying about duplicates. I
   want to use this to output a preliminary file that I can later submit
   to Last.fm.

The queue id counter resets at restart, but the queue is not yet
persisted either, so that is not a problem. Once I start persisting the
queue, I should persist the fresh id counter with it.