Remove type argument in favor of dynamic dispatch
It hurts a bit. I don't want to have to use dynamic dispatch, it is not
a zero-cost abstraction! But it does simplify things, and the few cycles
that these virtual calls cost are not going to be significant. Which is
not to say that they are unimportant, because the little bits stack up.

Initially I thought of having one MetaIndex interface, and an in-memory
implementation, and a disk-backed mmap implementation. But at this point
it is pretty clear that just an in-memory one is going to be sufficient,
and fit easily even for large collections on low-end devices. So
probably in the future I will remove the trait and just make the memory
metaindex the one. At that point the cost of the virtual calls will go
away again.