Log json on track start and complete
Save for the Musicbrainz track id, this is everything I need to submit a
Last.fm scrobble, plus the internal ids, just in case they turn out to
be useful. I plan to log this to a file next, and then I can later build
a scrobbler for that file, which I might integrate directly into the
daemon later, or I might keep it a separate service ... We'll see.