Add animation for opening the status bar
I intend for this to open up to a "now playing" or "overview" page,
which should show the currently playing track info, and the queued
tracks and history, but it should also have some more playback controls,
like pause and volume. I'm not entirely sure yet how that is going to
fit in, but let's see how it goes.