Begin working on new paned app layout
I have this huge amount of space on desktop, and I also have trouble
figuring out how to animate things on narrow viewports. So I think in
the end it will be better to divide the viewport up in panes, and let
those have their own scrollbars. Not sure how well this is going to work
out on Chrome on mobile with its changing viewport heigt, problems with
that were my original reason for moving away from nested parts with
their own scroll bars. I hope this will make it easier to fit the pieces
together.