Add script to convert Listenbrainz json to tsv
I thought I could do this with serde, but parsing nested fields into a
flat struct appears to be difficult (unlike the converse), and streaming
parsing is difficult too, I think it can be done with a visitor, but you
need a surrounding struct to add the annotation ... it all seems like
the wrong tool for what I want to do.

So instead, here are 5 lines of Python (plus some boilerplate) to
convert the Listenbrainz json into a tsv file, which should be a lot
easier to parse in Rust.