Match listens to track ids
This works surprisingly well for now. There are still lots of missing
tracks that should be present in the library (but which are not a match
due to casing differences, or feat. artist having moved out of the
title, and things like that). But it does match the majority of listens,
so I will just use this for now and fine-tune it later.