Make status bar switch back to album
Without this, on mobile there is no way (yet) to get back to the album
view, aside from the back button. If we navigate back to the album, then
from there you can use the new back button to get to the album list.

Also, it's nice to be able to jump to the current album, I did not have
that yet either.