Move album list to the right
I'm struggling a bit with the album view, on the one hand I really liked
the minimal page that I had before including the album view. On the
other hand, it wastes a lot of space, and it is very useful to be able
to switch between albums without the extra click.

However, including the album list disurbs the otherwise quiet album view
page. It has colorful album covers, and bold titles, it draws a lot of
attention to it.

I experimented a bit with fading it out except on hover, or sliding it
in/out. Fading works but is perhaps a bit cheesy, sliding works very
well to restore the minimalism of the page, but then you run into space
issues or alignment issues.

Moving the album list to the right does not solve any of these problems,
but I feel it does help to make the page look somewhat more quiet, the
list draws less attention to itself on the right. Maybe I should add the
fading as well, and then it might be okay.