Cap search results to keep the page responsive
If there are many results, creating the DOM nodes (and decoding the
thumbnails too, probably) is expensive. So expensive even, that it
delays typing. By capping, I get a smooth search-as-you-type experience,
however this no longer renders all results.

The proper solution is to add a "more ..." button for the case where we
do cap, or perhaps to do an infinite scroll like the album view, but
that would be a next step.