Authenticate with Last.fm API
I am now able to make an API call, however, my scrobble is rejected with
code 1 ("Artist was ignored"). I suppose this is because I have a feat.
artist in there, and Last.fm does not like to have these in artists. I
can try including a MusicBrainz id, perhaps that will help.