Write the play log as SQLite db, not as json file
I have the scrobbling script now, but I want a way to keep track of
which listens have been scrobbled already. I could do something with
moving the log file around, and hoping that we don't write when I do the
move, or posix advisory locks ... but then it is difficult to combine
the data again later, and it starts to become a pile of hacks ... so
let's use a database instead.

SQLite is the perfect tool for this:

 * It is easy to use from Rust (the "sqlite" crate exposes a pretty
   clean interface, there is not much to it, and it does not force
   unreasonable dependencies on to me).

 * It is easy to use from Python, built into the standard library.

 * It can handle multiple processes accessing the same database file
   concurrently.

 * It is embedded, so it doesn't require running an extra daemon, and it
   requires zero new configuration.

For now I roughly copied the play log format into the database. In the
future I might want to mirror the library itself in SQLite, in addition
to the in-memory data structures, so the play log can be normalized a
bit, which would also make it more usful for querying.

The schema of the database is semi-append-only: I plan to mutate rows
to add the completed time and scrobbled time, but aside from those, the
listens table is append-only. It could be made fully append-only by
recording completed and scrobble events in separate tables, with a
foreign key to the listens table, but potato potato, the data and
operations are the same, it is just a slightly different representation.
With the current schema that will have mutations, the mutations are very
well-behaved: only filling null values with something non-null, so you
get this lattice-like structure on rows where the state can only advance
until a row is in its final state (both completed and scrobbled). I
think that makes a mutation justified.