Reset prepared statement before reuse
So I compiled and ran the previous commit, and it all worked at the
first run, creating the database, inserting, updating the row ... it was
too good to be true. And indeed it was, inserting the second listen
failed. Turns out you have to reset the prepared statement before reuse.