Add tool to insert play log into SQLite database
I now built up a json play log over the past few weeks, and now that I
am moving to a database for the play log, I don't want to lose that
data. So add a script to insert the play log into the database.

This breaks the scrobble script temporarily. I will adapt it later to
use the SQLite database to figure out what to scrobble, and to mark the
tracks as scrobbled. For now, this is the way to migrate my data.