Add specialized album id hash table
I want to store thumbnails in memory to avoid disk access latency when
the disk has spun down for power saving. I could simply do

    thumbs: HashTable<AlbumId, Vec<u8>>

or

    thumbs: HashTable<AlbumId, Box<[u8]>>

And that would probably work well enough, but where's the fun in that?
It would put pressure on the allocator and waste precious memory to
store the slices separately, so let's fix that by concatenating them:

    thumb_data: Box<[u8]>
    thumbs: HashTable<AlbumId, (u32, u32)>

where these (u32, u32) pairs are index and offset into the data array.

Okay, nice, but we still have this general-purpose hash table. Now
Rust's hash table is pretty good, but it is general-purpose, it has to
be able to do things like deleting entries. I don't, I just want to
build this table once at startup, and then have the lookups be as fast
as possible. Also, album ids are taken from MusicBrainz UUIDs, so they
are already uniformly distributed, no need to hash them.

This is probably overkill, and it is questionable whether my hash table
is even better than the vectorized one in the standard library. But it
is a fun thing to get nerd sniped on, so I'm doing it anyway.