Tweak decode buffer memory management
I noticed when checking the logs that when the decoder wakes up, it has
just below 30 seconds of audio left, but usually around 40 MB of audio
data as well, which is far too much for 30 seconds, and takes up most of
the buffer room. This happens because I break the buffer up in parts,
and parts get discarded only after they have been fully played. When
decoding produces only a single part per file, it will be a large part,
so a lot of memory can be wasted on already-played samples, which means
the decoder has to wake up more. To counter this, produce parts of at
most ~10 MB per iteration. This way, we waste at most 10 MB on already-
played samples, and we can fill most of the buffer.

The cost is a bit more bookkeeping overhead, but I suspect that is
negligible, 10 MB parts is already quite large.