Implement intersection in search
This makes search a whole lot more useful: you can now search for
multiple words to nail down the query.

When trying this out, I did get results, but they did weird things:
"Murder Mi" would not match, but "Murder Mit" would match "Mitch
Murder". After putting prints all over the place, I discovered that I
reversed the comparisons while iterating the interseciton. Flipping
those then made it work as expected.

Ranking is not using all match metadata yet, but it seems to work pretty
okay already, probably because by the time you get to a second word,
there are few enough matches that they fit on the screen, such that
ranking is not as important any more.

There are a few more edge cases to tweak, in particular when a word
occurs twice in a title, or when a word occurs twice in the search
query, but I can tackle those later. For now this is already pretty
useful.