Handle repeated words in search query
I should really make a test for this too, but it is kind of tedious to
set up the data.

In any case, when you now search for "Much", it brings up "Much too
Much" and many others, and when you search for "Much Much", it narrows
it down to "Much to Much", so we match every query word only once to a
word in the title.