Improve Listenbrainz matching
There are still many many edge cases that don't get matched. E.g. feat.
artists that I moved out of the title into the artist, suffixes like
"bonus track" on the album title, feat. artists missing in the original
scrobble, etc.

I think a better bet would be to see if these listens include a
MusicBrainz id, and to match on that instead.