Add another point to the to do list
Currently I edit the tags for albums with multiple artists so one artist
is the primary one. You have to pick one either way if you follow an

  artist / album / track

file hierarchy. Still, it would be nice to support multiple artists as a
first-class feature. This would make the album show up when searching
for both, instead of making all the tracks pop up in search for one of
the two artists (because the other artist is still credited on each
track).