Handle errors from write_samples during playback
My daemon just died at the end of playing a track, with this error:

    thread 'playback' panicked at 'TODO: Failed to write samples.: Error("snd_pcm_mmap_commit", Sys(EPIPE))', src/playback.rs:262:15

Going by the sparse Alsa docs, I suspect the EPIPE is a buffer underrun.
The downside of Results is that I can't see which particular call this
one originated from; "write_samples", which is where the Err is returned
from, does handle error conditions, but not all of them. So I'm not
entirely sure what the right response is, but I suppose that just
continuing should work: if the device is in a bad state, "write_samples"
will recover (at least, it should). It might mean that some audio is
lost and there is a stutter, but if there is a buffer underrun, there is
a stutter anyway.

I don't know of a way to test this; I've been using Musium for half a
year now, and this is the first time I run into this case, so it is
extremely rare, and it would be difficult to reproduce. So I hope this
is an appropriate fix, but I have no way to confirm.