Use typography rather than icons
It's easier, and it also looks better, and takes up less space for small
screens (because the text can be small and wide, but icons would need to
be taller). The downside is that it may not be as immediately obvious,
icons can be seen at a glance, text you have to read. But for now this
will do.