Add endpoint to serve artist info
For now this only includes the artist name, and all of the albums. There
is some redundancy in the data, but I think that is fine, better than
needing bespoke parsers for every endpoint. Also, in the future I might
want to make a distinction between the album artist as text, the name of
an artist, and the artists (by id) of an album, in order to support
collaboration albums that have multiple artists.