Begin working on benchmark script
I want to measure how some changes in Claxon affect indexing
performance. Manually running "perf stat" and copy-pasting the result
into a table is getting tedious, let's automate this.

This will just be a thin wrapper that runs "perf stat" a number of times
and parser the results into a file that can then be loaded into e.g. R
for futher processing.

Alternatively, I might add some more SciPy stuff and run the statistical
tests from the script itself.

This commit adds step 1 of this: a way to call "perf stat" and parse the
result. Step 2 would be to do this in a loop and write the result to a
file.