Add comparison mode to benchmark script
This performs a Mann-Whitney U test on the before and after data. The
Mann-Whitney U test is great for benchmarks, because unlike the
Student's T test, it does not assume that the underlying distribution is
normal. A normal distribution is a questionable assumption for
benchmarks, because things like interrupts and mispredicts cause delays;
there is never a speedup over the minimal running time. If you model
those delays with a Beta distribution, the sum of many delays becomes a
Gamma distribution, and as the number of delays grows, it does tend
towards a normal distribution, but it's not clear to me how close this
is. Also, I spent a lot of time analyzing benchmark data, and even
assuming that delays follow a Beta distribution is already questionable.

The Mann-Whitney U test works under much weaker assumptions. The
trade-off is that it has less power for a given number of samples, if
the distribution *is* actually normal. I think this is fine, because
benchmarks are automated, we can just collect more samples.