Expose buffering status from API
This should later allow the webinterface to show a loader when the
player is blocked decoding (likely blocked on IO). Right now it already
advances the playback position, but then it hangs, and that is annoying.
I know that it is blocked waiting for the disk to spin up, but it's
still bad UX.

This also resolves a few to dos in how the queue is served.