Render a spinner when a track is buffering
This should give a bit better feedback when playing a track, but the
disk needs to spin up, so it takes like 15 seconds before it actually
starts playing. Rather than keeping the playhead in a fixed position 5
seconds out, we will now keep it at zero, hide the thumbnail, and show
a spinner instead. When the track starts playing, the spinner should
disappear, and the cover will unveal itself again.