Serve correct playback position when buffering
To compute the playback position in seconds, we need to know the sample
rate. But to know the sample rate, we need to have started decoding the
track. Previously the sample rate was only stored in blocks, but in case
of a buffer underrun, where we played all blocks, but the decoder can't
keep up producing new ones (for instance, because it is waiting for a
slow disk), that information was lost, so the playback position would
jump back to 0. Fix that by also storing the sample rate in the queued
track itself.