Try to prevent spinner from showing too eagerly
Now the spinner shows up correctly when the track is buffering, but what
happens now is that I enqueue a track, we show the thumbnail because
that is the happy path, but upon refresh it turns out that the player is
still buffering and the buffer is empty, so the track is blocked, and we
show the spinner. Then on the next refresh the spinner gets disabled
again, so we get this quick flash of a spinner. If it's that fast, we
could skip the spinner entirely anyway -- only start showing it if
buffering is going on for long enough.

What I hope to achieve is that in the case of the disk spinning,
enqueueing a track is fast enough that we don't have to show the spinner
loading animation, but when the disk is powered down and it has to spin
up first, then we start showing the spinner until the track starts
playing.