Retain "queued" dots after visiting album again
We mark tracks on the album with a "queued" visual style after queueing
the track, but this information was only in the DOM, after navigating
away from the album, and then returning to the album view, it was
impossible to see which tracks were queued. Fix that by taking the
queued tracks into account when rendering the album list.

This still does not *clear* the queued status of tracks in the album
view when the queue changes. It didn't do that previously either, that
remains to be done in future work.