Mark search input field as type=search
This first and foremost removes the annoying autofill bar in Chrome
mobile that covers part of the screen with a useless bar with only a few
icons in it.

Additionally, it turns the "enter" button on a mobile keyboard into a
search icon, and it also seems to add a cross in the input field to
clear it on Chrome mobile, which is nice, but unforunately not styled
as nicely as it could be.