Add fallback for album and artist navigation
If you open the app and have no previously visited album or artist, it
makes sense to go to that of the currently playing track, if there is a
currently playing track.

We can't yet go to the artist, because we don't know who the artist is,
but we can serve that from the API, because we do a lookup for the album
anyway.