Use track or album loudness based on context
If there are consecutive tracks from the same album, we want to use the
album loudness for loudness normalization, so that the differences
between tracks remain preserved as the master intended. But if the
preceding track and the next track are from a different album than the
current track, then we should normalize to track loudness, to achieve a
more uniform loudness overall.

This achieves that, by selecting the volume to use when we switch to a
track, based on the previous and next track at the time when we switch.
The queue can change afterwards, but we don't want to modify the volume
mid-track if the queue changes.