Prevent decoding too little per burst
Occasionally when I start a new track from an empty queue, I hear this
stuttering popping sound. I suspect that what is happening, is that the
decoder decodes a very tiny amount of samples, and those get played back
before the decoder gets a chance to decode more, leading to a buffer
underrun and then a pop. To mitigate this, put a lower bound on how much
we decode.

This sacrifices a bit of responsiveness for smoother playback.

But now that I am writing this, I realize it might also just be that the
decoder is plenty fast, but the disk is slow. Just after the disk spins
up, maybe readahead is not yet effective, and reads are slow, until the
kernel recognizes that we are doing a linear read through the file, and
starts reading larger chunks at once?

I'm not really sure what is causing the popping, but if this commit
resolves it, then it is clear what the cause was. If not, I'll revert
this commit.