Add bindings for sd_notify
When Musium runs under systemd, during startup, it still prints it
indexing progress to stdout with carriage returns. This is nice in a
terminal when watching live, but when running under systemd, it produces
this:

    musium[9930]:   library_path = /pool/music
    musium[9930]:   covers_path = /pool/volatile/covers
    musium[9930]:   data_path = /pool/volatile/musium
    musium[9930]:   audio_device = HDA Intel PCH
    musium[9930]:   audio_volume_control = Master
    musium[9930]: [16.7K blob data]

Note that "16.7K blob data" there -- that is not exactly helpful. And
because this is disk-bound, this can take a minute or two. It would be
nice to see progress. Systemd has a mechanism to report progress. This
adds bindings for that. The next step would be to replace status
reporting.