Add an interface for status reporting
Right now it is only implemented for io::Write, which should be an
StdoutWriteLock, so nothing really changes in the behavior. (The message
is a little different though.) My next step will be to implement this in
a way that makes progress updates go to systemd instead of stdout, while
issues still go to stdout.