Move all API endpoints behind /api path
This way, if I want to have human-readable urls, they can be served from
the app server without getting in the way of the API.

This can later be solved to solve a minor cosmetic issue: the fact that
I had to use query parameters for navigation, because otherwise reload
would be broken. I would prefer to have urls like /album/abc and
/artist/abc instead of /?album=abc and /?artist=abc.

Credit goes to Mattermost for this idea; it also has a single server for
both web and the API, and all API calls are behind the /api path. Simple
but effective.