Replace apply-transition hack with more robust fix
In Musium I rely heavily on css transitions, and adding and removing
classes to trigger them. But if you add a class and then remove it, if
no style recalc happens in between, then it's a no-op, and you get no
transition.

Previously I worked around this by adding a small delay. It worked,
mostly, but I occasionally noticed that transitions were missing. I
suspected this is because the delay was not long enough, and the style
changes were batched into one recalc, becoming a no-op.

Instead of using a delay, we can consistently force a style
recalc by calling a function that requires a style recalc, such as
getBoundingClientRect. So far I haven't observed any dropped
transitions now.