Step towards avoiding stutter when opening album
I discovered the cause of the occasional stutter: after an image loads
and is decoded, it needs to be repainted, which can take a long time
(~200ms up to 2 full seconds in bad cases) in Chromium for large images
that need to be heavily downsampled.

When the image wasn't loaded before the album view started opening, but
finished loading during the transition, that would stall the main
thread, and the transition would stutter.

To mitigate this, I want to:

 * Kick off loading the cover instantly.
 * Wait very briefly before the transition, add the cover if it's
   available.
 * If not, only add it *after* the transition.

This leaves the album view in a broken state, because we still need to
call the advance functions in the appropriate places. I will address
that in a follow-up commit.