Add the track list and full-res cover back
Now instead of adding them immediately when they are available, we have
control over how they are loaded, which should make transition
animations smoother.

Indeed, they do, but there are new cases now that are annoying:

 * Loading the full-res cover can sometimes cause a visual "pop".
   Previously, the majority of the time it would happen during the
   transition, so it would be much less apparent.

   But in any case, this is mostly caused by cover art not using sRGB
   but my thumdnails being converted to sRGB, so perhaps this can be
   mitigated by improving the thumbnailer.

 * Usually the track list loads in time before the transition, but when
   it doesn't, it now pops into place after. It needs a transition
   itself for that case.

 * Clicking an album is less responsive because we introduced the little
   wait period. Previously, starting the transition already allowed for
   some slack in the loading time, if things popped in later during the
   transition, that was usually fine.

So, overall I'm not sure whether this is an improvement. It makes things
smoother in some cases, but it regresses in others. Perhaps a good
middle ground would be to allow the track list to pop in, but not the
album art. Something else that might help is to already start loading
the image on mousedown/touch, instead of only starting on navigate.