Open "now playing" initially when queue is nonempty
I noticed that I often want to check what the current song is when
something is playing, so let's open right at that page if something is
playing. It makes more sense to me either way like this.

This does still show the transition, I'll have to make an exception for
that so it loads instantly instead.