Split database logic into module, add files table
For a long time, indexing at startup has been bothering me. Because it
is IO-bound, it takes about two minutes for me. After this the files are
in the page cache, so it does not impede debugging too much, but still,
it's a nuisance. Also when I add something to my collection, I restart
the daemon, and it takes a few minutes.

So it would be nice if we could persist the previous index, then figure
out which files changed, and update the index. Then we should get near-
instant restarts if no files changed, and even when adding a few albums,
that should only take a few dozen milliseconds.

I struggled for a long time with how to represent this. SQLite was
probably better than a custom binary format, as I was already using it
anyway, and it takes care of transactions and such. But what to put in
SQLite? At first I thought: “three tables: tracks, albums, artists”.
Saving an index to that should be easy. Rebuilding would maybe be a bit
unnaturaly, but we can still flatten the three tables in a query to
produce all data per track, and feed that into the indexer. The tricky
thing would be invalidation: how do you know the effects of changes in
one file? It could make an album or even artist title inconsistent. What
if a file gets added to an album, or removed? I'm probably worrying too
much about a non-issue, because in practice I only ever add files, and
in the rare cases that I do change something, it would be fine to just
delete the saved data and do a full index, it still only takes a few
minutes.

But then I realized: I shouldn't store the structured data in SQLite. I
should just store the raw data from the files in there. Musium has its
own index data structures, it wouldn't use SQLite for relational queries
anyway. But the nice thing is that now invalidation is easier, because
the rows in the database map 1:1 to the files on disk, so we can do a
set difference, delete any modified or deleted files from the database,
and then add any files that are not present. And subsequent indexing
doesn't change much, it still takes the same inputs, only now from
SQLite instead of directly from the files. We can even split this out
into a separate process: something could fill or update the database in
the background, and Musium can already start and begin serving, and
reload the new tracks later. It will nicely separate the concerns of
collecting the data, from building the index.